Testing AI
There are 2 different scenes in the imported package of the AI InWorld. The 2D scene offers a textbox that has an AI response and AI audio that repeats what the text says. It is pretty user-friendly and reminds me of one of those website chatbots to help you find what you need. In the 3D scene, you are standing in front of the AI and talking to it directly. They provided a model and the model even changes facial expressions and gestures depending on your responses. They also have a “Ready Player Me” avatar in one of the scenes and to navigate in this scene they have set up WASD to move around. They even have a recording option which would help the AI understand our language and respond to it accordingly.